:ambulance: (network) Finally fix the network login process

+++ protocol/src/main/kotlin/me/hexalite/liteware/protocol/annotations/MinecraftPacketInfo.kt
@@ -2,4 +2,4 @@ package me.hexalite.liteware.protocol.annotations
 
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.RUNTIME)
-annotation class MinecraftPacketInfo(val id: Int)
+annotation class MinecraftPacketInfo(val id: Byte)
